+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Group service: implement loadGroupEventMemberSelection [#1519](https://github.com/rokwire/illinois-app/issues/1519).
 ### Fixed
 - Fixed crash on activity destroy in Android native side (#50).
+- Image rotation before upload [#58](https://github.com/rokwire/app-flutter-plugin/issues/58).
 ### Added
 - Get, Create, Update and Delete user's profile picture [#53](https://github.com/rokwire/app-flutter-plugin/issues/53).